Fix warnings reported by MSVC 2013
Reviewed-on: http://review.couchbase.org/45800
Tested-by: buildbot <build@couchbase.com>

+++ src/cJSON.c
@@ -495,7 +495,7 @@ static char *print_array(cJSON *item, int depth, int fmt)
         ret = print_value(child, depth + 1, fmt);
         entries[i++] = ret;
         if (ret) {
-            len += strlen(ret) + 2 + (fmt ? 1 : 0);
+            len += (int)strlen(ret) + 2 + (fmt ? 1 : 0);
         } else {
             fail = 1;
         }
@@ -635,7 +635,7 @@ static char *print_object(cJSON *item, int depth, int fmt)
         names[i] = str = print_string_ptr(child->string);
         entries[i++] = ret = print_value(child, depth, fmt);
         if (str && ret) {
-            len += strlen(ret) + strlen(str) + 2 + (fmt ? 2 + depth : 0);
+            len += (int)strlen(ret) + (int)strlen(str) + 2 + (fmt ? 2 + depth : 0);
         } else {
             fail = 1;
         }
